add components.py
